projects/fun_search/temp_requirements.txt
accelerate==0.26.1
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
crashtest==0.4.1
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.5.14
distlib==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
dulwich==0.21.7
exceptiongroup==1.2.0
Farama-Notifications==0.0.4
fastjsonschema==2.19.1
filelock
Flask==3.0.0
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.27.0
google-cloud-aiplatform==1.40.0
google-cloud-bigquery==3.17.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
gymnasium==0.29.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
hydra-core==1.3.2
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
keyring==24.3.0
kiwisolver==1.4.5
langchain==0.1.2
langchain-community==0.0.14
langchain-core==0.1.14
langchainplus-sdk==0.0.20
langsmith==0.0.83
litellm==1.19.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
minigrid==2.3.1
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numexpr==2.8.8
numpy==1.26.3
omegaconf==2.3.0
openai==1.5.0
openapi-schema-pydantic==1.2.4
packaging==23.2
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==3.11.0
plotly==5.18.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
poetry2setup==1.1.0
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pygame==2.5.2
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
SecretStorage==3.3.3
sentry-sdk==1.39.1
setproctitle==1.3.3
shapely==2.0.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.24
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.66.1
transformers==4.36.2
trove-classifiers==2023.11.29
typing-inspect==0.9.0
typing_extensions==4.5.0
urllib3==2.1.0
virtualenv==20.25.0
wandb==0.16.1
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
